+++ cmd/server/server.go
@@ -2,10 +2,9 @@ package main
 
 import (
 	"context"
-	"github/JustGopher/Gotaxy/internal/tunnel/serverCore/pool"
-
-	"github/JustGopher/Gotaxy/internal/tunnel/serverCore/global"
-	"github/JustGopher/Gotaxy/internal/tunnel/serverCore/shell"
+	"github/JustGopher/Gotaxy/internal/global"
+	"github/JustGopher/Gotaxy/internal/pool"
+	shell2 "github/JustGopher/Gotaxy/internal/shell"
 )
 
 func main() {
@@ -15,7 +14,10 @@ func main() {
 	global.ConnPool.Set("9080", "127.0.0.1:8080")
 	global.ConnPool.Set("9081", "127.0.0.1:8081")
 
-	sh := shell.New()
-	shell.RegisterCMD(sh)
+	global.ListenPort = "9000"
+	global.ServerIP = "127.0.0.1"
+
+	sh := shell2.New()
+	shell2.RegisterCMD(sh)
 	sh.Run()
 }
