--- internal/tunnel/serverCore/global/global.go
@@ -1,12 +1,0 @@
-package global
-
-import (
-	"context"
-	"github/JustGopher/Gotaxy/internal/tunnel/serverCore/pool"
-)
-
-var (
-	Ctx      context.Context
-	Cancel   context.CancelFunc
-	ConnPool *pool.Pool
-)
