--- pkg/globals/global.go
@@ -1,7 +1,0 @@
-package globals
-
-import "go.uber.org/zap"
-
-var (
-	Log *zap.SugaredLogger
-)
